+++ proto/src/prost/cosmos.base.kv.v1beta1.rs
@@ -2,13 +2,13 @@
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Pairs {
     #[prost(message, repeated, tag="1")]
-    pub pairs: ::std::vec::Vec<Pair>,
+    pub pairs: ::prost::alloc::vec::Vec<Pair>,
 }
 /// Pair defines a key/value bytes tuple.
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Pair {
-    #[prost(bytes, tag="1")]
-    pub key: std::vec::Vec<u8>,
-    #[prost(bytes, tag="2")]
-    pub value: std::vec::Vec<u8>,
+    #[prost(bytes="vec", tag="1")]
+    pub key: ::prost::alloc::vec::Vec<u8>,
+    #[prost(bytes="vec", tag="2")]
+    pub value: ::prost::alloc::vec::Vec<u8>,
 }